get latest closed pnl append to CPNL csv

+++ nebixbm/command_center/tools/Tracer.py
@@ -103,9 +103,14 @@ class Tracer:
                 "Leverage",
                 "CreatedAt",
             ]
-            with open(self.cpnl_tracer_path, "w", newline="") as csv_file:
-                writer = csv.writer(csv_file)
-                writer.writerow(header)
+            if not os.path.isfile(self.cpnl_tracer_path):
+                with open(self.cpnl_tracer_path, "w", newline="") as csv_file:
+                    writer = csv.writer(csv_file)
+                    writer.writerow(header)
+                self.logger.info("New CPNL csv file has created.")
+            else:
+                self.logger.warning("CPNL csv file already exists. "
+                                    "Further datum will be appended.")
 
             self.logger.debug(
                 "Tracer initialized.\n" +
